+++ Changelog.md
@@ -8,6 +8,7 @@
 
 - ### Improvements
 
+  - [#492](../../issues/492) - Improve Property List (sort by multiple columns + add property type as column) (thanks @todor-dk)
   - [#504](../../issues/504) - Allow to scope the treeview to an element (thanks @Laniusexcubitor)
   - [#508](../../issues/508) - Persist UseRegex in settings (thanks @corvinsz)
   - Improving clear filter handling and keyboard focus for property edits